Continued work on bringing Fiber based Task Scheduler to macOS: - changed "aligned" to "alignTo" and "aligndefault" to "alignToDefault", to ensure that there is no collision with macOS/clang preprocessor/defines - removed query of thread sleeping state, added ability for threads to sleep till given time, or for given time - added ability to name threads for debugging purposes (need to be enabled on Windows, once migrated to VS2017 and latest Windows SDK) - added proper query for amount of physical and logical cores on macOS - added runtime check, verifying Mutex allocation on cacheline boundary - renamed "OSXInterface" to "macInput" to properly describe input interface - introduced unified engine initialization and deinitialization - properly passing main thread to AppDelegate Windows related changes need to follow this commit.

+++ public/include/memory/workStealingDeque.h
@@ -182,8 +182,8 @@ namespace en
          top(0),
          bottom(0),
          buffer(nullptr),
-         size(roundUp(sizeof(T) * capacity, 4096)),
-         maxSize(roundUp(sizeof(T) * maxCapacity, 4096)),
+         size(roundUp(static_cast<uint64>(sizeof(T)) * capacity, static_cast<uint64>(4096))),
+         maxSize(roundUp(static_cast<uint64>(sizeof(T)) * maxCapacity, static_cast<uint64>(4096))),
          doubleSizeUntil(_doubleSizeUntil)
    {
    // Managed elements cannot exceed pointer size (max 8 bytes), to ensure that
